Create quantum_number_generation.py
quantum_coin_generation or quantum_number_generation
In this example:
We connect a node (Alice) and create a fresh qubit then apply a Hadamard gate. After measuring the qubit, we repeat the process 10 times(in the for loop)and finally  we create an empty list in order to saving the quantum numbers or coins and the quantum numbers/coins are printed in the screen
Note: Some resources  say quantum random number generation for the process and some resources say quantum coin generation. In my opinion quantum coin generation is a quantum money and I think it is more different process although it has some similar features. To prevent confussion, I explained here